Implement Full Page Printing for Haiku Port
Switched from using viewport snapshotting to PrintContext-based rendering
to support full page printing.

* Source/WebKit/UIProcess/API/haiku/PageClientImplHaiku.cpp:
(WebKit::AsyncPrinter::snapshotNextPage): Use drawRectToImage instead of takeSnapshot. Added PrintInfo.h include. Explicitly cast float dimensions to int.
* Source/WebKit/WebProcess/WebPage/WebPage.h: Enable drawRectToImage for Haiku.
* Source/WebKit/WebProcess/WebPage/WebPage.messages.in: Enable DrawRectToImage message for Haiku.
* Source/WebKit/WebProcess/WebPage/haiku/WebPageHaiku.cpp:
(WebKit::WebPage::drawRectToImage): Implemented using ShareableBitmap and PrintContext.

+++ Source/WebKit/UIProcess/API/haiku/PageClientImplHaiku.cpp
@@ -470,7 +470,7 @@ private:
         printInfo.pageSetupScaleFactor = 1.0;
         printInfo.availablePaperWidth = printableRect.Width();
         printInfo.availablePaperHeight = printableRect.Height();
-        printInfo.rect = IntRect(0, 0, printableRect.Width(), printableRect.Height());
+        printInfo.rect = IntRect(0, 0, (int)printableRect.Width(), (int)printableRect.Height());
 
         m_page.drawRectToImage(m_frame.frameID(), printInfo, rect, snapshotSize, [this, protectedThis = Ref { *this }](std::optional<ShareableBitmap::Handle>&& imageHandle) {
             if (imageHandle) {
